feat(#317): make processing more generic
Even the CUDA-code is disabled at currently, the
basic processing workflow between the cpu- and
cuda-case was extremely different. This was fixed
by defining a more generic way based on the worker-
thread apporach of the cpu-case. This will make it
eaiser to implement a new CUDA-support.
