+++ src/hanami/src/core/processing/cuda/cuda_worker_thread.h
@@ -1,5 +1,5 @@
 /**
- * @file        cuda_host.h
+ * @file        cuda_worker_thread.h
  *
  * @author      Tobias Anker <tobias.anker@kitsunemimi.moe>
  *
@@ -20,35 +20,29 @@
  *      limitations under the License.
  */
 
-#ifndef CUDAHOST_H
-#define CUDAHOST_H
+#ifndef CUDAWORKERTHREAD_H
+#define CUDAWORKERTHREAD_H
 
-#include <core/processing/logical_host.h>
+#include <core/processing/cuda/cuda_host.h>
+#include <hanami_common/threading/thread.h>
 
-#include <mutex>
+class Cluster;
 
-class CudaHost : public LogicalHost
+class CudaWorkerThread : public Hanami::Thread
 {
    public:
-    CudaHost(const uint32_t localId);
-    ~CudaHost();
-
-    void addClusterToHost(Cluster* cluster);
-
-    bool moveCluster(Cluster* cluster);
-    void syncWithHost(Cluster* cluster);
-    void removeCluster(Cluster* cluster);
-
-    std::mutex cudaMutex;
+    CudaWorkerThread(CudaHost* host);
+    ~CudaWorkerThread();
 
    protected:
-    bool initWorkerThreads();
+    void run();
 
    private:
-    void initBuffer();
+    void trainClusterForward(Cluster* cluster);
+    void trainClusterBackward(Cluster* cluster);
+    void requestCluster(Cluster* cluster);
 
-    std::atomic_flag m_queue_lock = ATOMIC_FLAG_INIT;
-    std::deque<Cluster*> m_clusterQueue;
+    CudaHost* m_host = nullptr;
 };
 
-#endif  // CUDAHOST_H
+#endif  // CUDAWORKERTHREAD_H
